test auto (#27)
* test

* terraform-docs: automated action

* test

* test

* Add code to convert an image's color using OpenAI

This commit adds code to convert the color of an image using OpenAI. The script imports the necessary libraries, converts a PNG RGB image to PNG RGBA, crops it, and saves it as a new file. It then uses OpenAI to create a variation of the edited image with a different color for the house. Finally, it prints out the URL of the new image.

* Add environment variable for OpenAI API key

This commit adds an import statement for the os module and replaces the hardcoded OpenAI API key with an environment variable. This change allows users to set their own API key without modifying the code.

* autopep8 action fixes (#28)

Co-authored-by: cailen <cailen@users.noreply.github.com>

* auto assign prs

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: cailen <cailen@users.noreply.github.com>